+++ CHANGELOG.md
@@ -44,6 +44,7 @@ applications:
  * [`PULL-180`](https://github.com/thiagoesteves/deployex/pull/180) Moving deploy timeouts to be handle by application and not globally
  * [`PULL-182`](https://github.com/thiagoesteves/deployex/pull/182) Adding app config information in the UI/UX and full restart button for all apps
  * [`PULL-186`](https://github.com/thiagoesteves/deployex/pull/186) Adding host uptime to UI/UX
+ * [`PULL-187`](https://github.com/thiagoesteves/deployex/pull/187) Update DeployEx to elixir 1.19.3
 
 ## 0.7.3 🚀 (2025-10-28)